Restored old update

+++ xbrl/__init__.py
@@ -1,2 +1,2 @@
 import torch
-TORCH_FLOAT = torch.float32
+TORCH_FLOAT = torch.float64
